[l10n] Tidy generated resource files
Add the automatic changes that Visual Studio makes when saving the
resource files so there won't be any need to worry about them getting
included unintentionally as part of an unrelated future change to the
`.resx` file.

Other changes:

Remove some space and newline characters from the `AAPTManifestError`
resource.  I think what happened was that the text editor auto-indented
the content of that resource in the original commit 70790837.  Since the
`.resx` XML parser preserves whitespace, those spaces will carry over
into the final MSBuild error message that the app developer sees, which
I think was not intended in this case.