[Xamarin.Android.Build.Tasks] Dont copy assets for Library Projects.
We currently copy all the assets for all Library Projects
into one main directory. This can slow down build times
if the assets are large. So what we should do is use those
assets in place.

This will also allow us to consume .nuget assets in place as
well sometime in the future.