[spec] for @(AndroidLibrary) and avoiding EmbeddedResource
Context: https://github.com/xamarin/xamarin-android/issues/2441

See `OneDotNetEmbeddedResources.md` for the new spec for
`@(AndroidLibrary)` and removal of `EmbeddedResource` usage in
Xamarin.Android class libraries.

Future proposals can be commited to `Documentation/proposals`. When
implemented, the proposal can be reworked and moved to
`Documentation/guides` if it is still useful.