[build] consolidate global.json for Microsoft.Build.NoTargets (#5104)
Remove `build-tools/create-packs/global.json` and have a single
`global.json` file in the root.  This allows us to remove the version
number for `Sdk="Microsoft.Build.NoTargets"` in several files.

We weren't able to do this in commit 079197ac because the extra
`global.json` file that took precedence over the one in the root.