[One .NET] fix <Import/> ordering
We were setting several MSBuild property defaults in
`Microsoft.Android.Sdk.props`. The problem with this, is that it
doesn't enable the properties to be overridden in a .csproj file at
all.

The import ordering should be:

1. `Microsoft.Android.Sdk.props` - only contains properties that
   cannot be overridden, or properties that the `.csproj` might need
   to rely on.
2. The `.csproj` file
3. `Microsoft.Android.Sdk.targets`
  a. `Microsoft.Android.Sdk.DefaultProperties.props`
  b. `Microsoft.NET.Sdk/Sdk.targets`
4. Any other `.targets` files where we have functionality split up.

I took the opportunity to reorganize/cleanup a few other things:

* `@(ProjectCapability)` should go in
  `Microsoft.Android.Sdk.DefaultItems.targets`.
* Lots of usage of `$(MSBuildThisFileDirectory)` that wasn't needed.
* Sorted `<Import/>` alphabetically on the ones that ordering doesn't
  matter.

After these changes, I found that `XASdkProject` wasn't properly
building for `Release` configurations. We needed the same workaround
to set `$(Configuration)` in `Directory.Build.props`. To do this I
renamed `SetExtraNuGetConfigSources` to `UseDotNet`.