[C++] Add support for native code static checkers and sanitizers
We now support the [`clang-tidy`][0] static analysis of the C++ code
and the [`ASAN`][1] and [`UBSAN`][2] clang sanitizers on the runtime to find issues
that cannot be detected by static analysis. All of the tools are shipped
with the Android NDK which currently ships the clang 9 toolchain.

All of the tools require separate builds of the native code.
`clang-tidy` builds the entire codebase to generate an AST and then
analyze it, reporting possible errors but it does not output any
binaries.

The two runtime sanitizers, however, instrument the generated code in
numerous ways thus making it much (up to 3x) slower to execute.
Therefore, we now build our native code 4 additional times - for each
sanitizer in both `Debug` and `Release` modes. The output binaries have
the `-checked+SANITIZER_NAME` infix added to their name.

The instrumented binaries will not be currently distributed to the end
users since they are intended for internal use.

Also added are two [`wrap.sh`][3] scripts, one for `ASAN` and the other
for `UBSAN` and can be found in the `build-tools/wrap.sh/` directory.

The instrumented binaries can then be placed in the APK instead of the
standard binaries by adding the following to the project file:

```xml
<PropertyGroup>
  <AndroidIncludeWrapSh Condition=" '$(UseASAN)' != '' Or '$(UseUBSAN)' != '' ">true</AndroidIncludeWrapSh>
  <_AndroidCheckedBuild Condition=" '$(UseASAN)' != '' ">asan</_AndroidCheckedBuild>
  <_AndroidCheckedBuild Condition=" '$(UseUBSAN)' != '' ">ubsan</_AndroidCheckedBuild>
  <_ASANScript>../path/to/build-scripts/wrap.sh/asan.sh</_ASANScript>
	<_UBSANScript>../path/to/build-scripts/wrap.sh/ubsan.sh</_UBSANScript>
</PropertyGroup>

<ItemGroup>
  <AndroidNativeLibrary Include="$(_ASANScript)" Condition=" '$(UseASAN)' != '' ">
    <Link>lib\arm64-v8a\wrap.sh</Link>
  </AndroidNativeLibrary>
  <AndroidNativeLibrary Include="$(_ASANScript)" Condition=" '$(UseASAN)' != '' ">
    <Link>lib\armeabi-v7a\wrap.sh</Link>
  </AndroidNativeLibrary>
  <AndroidNativeLibrary Include="$(_ASANScript)" Condition=" '$(UseASAN)' != '' ">
    <Link>lib\x86\wrap.sh</Link>
  </AndroidNativeLibrary>
  <AndroidNativeLibrary Include="$(_ASANScript)" Condition=" '$(UseASAN)' != '' ">
    <Link>lib\x86_64\wrap.sh</Link>
  </AndroidNativeLibrary>

  <AndroidNativeLibrary Include="$(_UBSANScript)" Condition=" '$(UseUBSAN)' != '' ">
    <Link>lib\arm64-v8a\wrap.sh</Link>
  </AndroidNativeLibrary>
  <AndroidNativeLibrary Include="$(_UBSANScript)" Condition=" '$(UseUBSAN)' != '' ">
    <Link>lib\armeabi-v7a\wrap.sh</Link>
  </AndroidNativeLibrary>
  <AndroidNativeLibrary Include="$(_UBSANScript)" Condition=" '$(UseUBSAN)' != '' ">
    <Link>lib\x86\wrap.sh</Link>
  </AndroidNativeLibrary>
  <AndroidNativeLibrary Include="$(_UBSANScript)" Condition=" '$(UseUBSAN)' != '' ">
    <Link>lib\x86_64\wrap.sh</Link>
  </AndroidNativeLibrary>
</ItemGroup>
```

The `Link` metadata value **must** be valid path to the native library
directory inside the APK since the wrapper scripts are
achitecture-specific and are expected by the Android loader to be found in
architecture-specific library directories.

The checked build is enabled by setting the `$(_AndroidCheckedBuild)`
MSBuild property to either `asan` or `ubsan` value.

[0]: https://releases.llvm.org/9.0.0/tools/clang/tools/extra/docs/clang-tidy/
[1]: https://releases.llvm.org/9.0.0/tools/clang/docs/AddressSanitizer.html
[2]: https://releases.llvm.org/9.0.0/tools/clang/docs/UndefinedBehaviorSanitizer.html
[3]: https://developer.android.com/ndk/guides/wrap-script