[tests] Fix process output collection
The `BaseTest.RunProcess` is redirecting the standard and error output,
so that it can be captured. It didn't work, because it was not using the
output related events and so if there was enough output, we got dead
lock and timeout as a result.

Context: https://docs.microsoft.com/en-us/dotnet/api/system.diagnostics.process.standarderror

Example behaviour before the fix - notice the timeout:

    RunProcess: C:\Users\rodo\android-toolchain\sdk\platform-tools\adb.exe  logcat -d > "C:\Users\rodo\git\xa-master\bin\TestDebug\temp\otNetInstallAndRunTrueFalse\logcat-failed.log"
    RunProcess timed out: C:\Users\rodo\android-toolchain\sdk\platform-tools\adb.exe  logcat -d > "C:\Users\rodo\git\xa-master\bin\TestDbug\temp\DotNetInstallAndRunTrueFalse\logcat-failed.log"