[Xamarin.Android.Build.Tasks] Guard against only having 1 processor. (#5134)
Fixes: https://dev.azure.com/DevDiv/DevDiv/_workitems/edit/1216790

We had this very weird situation where the `<Aapt2Compile/>` task
would hang indefinitely.  This did not happen on every machine,
just on a particular customer's setup.

It turns out that `Environment.ProcessorCount` was returning 1.
This would cause the following code to return a value of 0.

	Math.Min(Environment.ProcessorCount - 1, DefaultMaxAapt2Daemons);

We did not have any protection in place to guard against getting a 0
from this code.  As a result we tried to spawn 0 instances of `aapt2`,
which would result in the `<Aapt2Compile/>` task waiting for jobs
that were never started, and thus could never complete.

Ensure that `Aapt2.Execute()` always uses *at least* one CPU.