[WIP, xaprepare] Generate portions of the test suite files
This commit attempts to organize the currently spread-out landscape of
Xamarin.Android tests so that there is only a single location where all
the tests can be listed, configured, added, removed etc. The idea is to
leave the current infrastructure (the various test .targets files) in
place, but generate portions of them (e.g. test-specific .projitem
files) based on description of tests.  The description can be found in
xaprepare's `ConfigAndData` directory, in the `Tests.cs` file. This
will be, ideally, the only location where one needs to make changes in
order to affect what tests, where and how are ran throughout
Xamarin.Android.

A new step, `Step_PrepareTests`, is added to xaprepare which takes care
of the actual generation.

Currently, the following test types are supported:

 * APK
 * Host Unit Tests