[One .NET] deprecate support for DebugType=Full or DebugType=PdbOnly
In .NET 5+ `DebugType=full` or `DebugType=pdbonly` does not appear to
be supported. These silently fall back to `portable`.

Testing with .NET 5 rc2:

* `dotnet new console`
* Add `<DebugType>full</DebugType>` or `<DebugType>pdbonly</DebugType>`
* `dotnet build` results in portable `.pdb` files on both Windows and
  Mac.

This is good news, as we've been wanting to deprecate the other symbol
formats *anyway*.

1. Emit a XA0125 warning for `full` or `pdbonly` symbol files for
   "legacy" Xamarin.Android. Move any related MSBuild targets to
   `Xamarin.Android.Legacy.targets`.

2. In Microsoft.Android.Sdk, we only use the symbol files returned by
   the `ComputeFilesToPublish` MSBuild target from the dotnet/sdk. The
   old symbol file formats appear to be ignored.

3. We no longer need to ship `pdb2mdb` in Microsoft.Android.Sdk.

I updated our MSBuild tests to default to using `DebugType=portable`.
I also updated a couple tests to rely on the new default value.