[Xamarin.Android.Build.Tests] Import tests from legacy definition (#5159)
Context: https://github.com/xamarin/QualityAssurance/pull/3648

Bring in additional Install + Run tests for mono-symbolicate, the
profiler, and the linker.  In the past, a few of these linker related
test scenarios would fail with a SIGSEGV which is why they exist in a
stand alone test case rather than being added to a unit test app.