[Xamarin.Android.Build.Tasks] remove acw-map.txt usage in <ConvertResourcesCases/>
The `<ConvertResourcesCases/>` loads the `acw-map.txt` file and passes
the dictionary throughout several methods in `AndroidResource`.
However, this dictionary is completely unused throughout?

At the time `<ConvertResourcesCases/>` runs (before `Build`),
`acw-map.txt` does not even exist. It is created by the
`<GenerateJavaStubs/>` MSBuild task, that runs after `Build`.

I was working on a performance improvement, where I was attempting to
use `acw-map.txt` -- unfortunately an empty dictionary.

To eliminate confusion and clean things up, I removed all usage of
`acw-map.txt` in the `<ConvertResourcesCases/>` MSBuild task and
`AndroidResource` class. I was not able to see a performance
improvement with this change, as it is likely too small to notice.