[VSAndroidApp] VS Android App template sample to measure apk size (#4793)
Add a clean project created in Visual Studio using the Android App
template.  The main idea here is to have clean project without any
changes, so that it can be easily updated with newer versions when
needed.  This can be done by replacing the `samples/VSAndroidApp`
directory with newer project generated in VS from the same template.

The project itself is located in the `samples` directory and is used
through the proxy project `VSAndroidAppProxy.csproj`, which makes it
possible to build it within the xamarin-android repo.

VSAndroidApp is also run during the **Smoke Tests** >
**APK Instrumentation** job on Azure, and the apk size is compared
using `apkdiff` against the reference file
`tests/apk-sizes-reference/com.companyname.vsandroidapp-Signed-Release.apkdesc`.