[Xamarin.Android.Build.Tasks] <ConvertResourcesCases/> & acw-map.txt (#5184)
The `<ConvertResourcesCases/>` task loads the `acw-map.txt` file and
passes the dictionary throughout several methods within
`AndroidResource`.

However, this dictionary is completely unused throughout?

Since 1886e6f1, the logic using `acw-map.txt` was moved to the
`<ConvertCustomView/>` task.  We no longer need to use this file in
the `<ConvertResourcesCases/>` task.

At the time `<ConvertResourcesCases/>` runs (before `Build`),
`acw-map.txt` does not even exist.  It is created by the
`<GenerateJavaStubs/>` MSBuild task, that runs after `Build`.

I was working on a performance improvement, where I was attempting to
use `acw-map.txt` -- unfortunately an empty dictionary.

To eliminate confusion and clean things up, I removed all usage of
`acw-map.txt` in the `<ConvertResourcesCases/>` MSBuild task and
`AndroidResource` class.  I was not able to see a performance
improvement with this change, as it is likely too small to notice.